+++ db/sql/create_tables.sql
@@ -15,7 +15,7 @@ CREATE TABLE exercises (
 CREATE TABLE workouts (
    id INTEGER PRIMARY KEY AUTOINCREMENT,
    name TEXT NOT NULL UNIQUE,
-   author TEXT NOT NULL,
+   user_id INTEGER NOT NULL,
    description TEXT NOT NULL,
    created TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
@@ -30,14 +30,21 @@ CREATE TABLE workout_exercises (
 
 CREATE TABLE exercise_likes (
    id INTEGER PRIMARY KEY AUTOINCREMENT,
-   username TEXT NOT NULL,
+   user_id INTEGER NOT NULL,
    exercise TEXT NOT NULL,
    created TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
 
 CREATE TABLE workout_likes (
    id INTEGER PRIMARY KEY AUTOINCREMENT,
-   username TEXT NOT NULL,
+   user_id INTEGER NOT NULL,
    workout TEXT NOT NULL,
    created TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP
 );
+
+--CREATE TABLE users (
+--   id INTEGER PRIMARY KEY AUTOINCREMENT,
+--   email TEXT NOT NULL UNIQUE,
+--   password TEXT NOT NULL,
+--   name TEXT NOT NULL
+--);
