+++ templates/add_exercise_to_workout.html
@@ -9,7 +9,7 @@
 <h2>Exercise List</h2>
 {% if selected_exercises %}
 {% for selected_exercise in selected_exercises %}
-    <a href="{{ url_for('exercise', exercise_id=selected_exercise['id']) }}">{{selected_exercise['name']}} </a>
+    <a href="{{ url_for('main.exercise', exercise_id=selected_exercise['id']) }}">{{selected_exercise['name']}} </a>
     <a>Sets: {{ selected_exercise['num_sets'] }}</a>
     <a>Reps: {{ selected_exercise['num_reps'] }}</a>
     <br>
@@ -20,7 +20,7 @@
 
 <h2>Add Exercises</h2>
 
-<form method="post" action="{{ url_for('add_exercise_to_workout', workout_id=workout['id']) }}">
+<form method="post" action="{{ url_for('main.add_exercise_to_workout', workout_id=workout['id']) }}">
 
     <div class="form-group">
         <select name=exercise class="form-control">
