+++ templates/base.html
@@ -12,30 +12,40 @@
 </head>
 <body>
 <nav class="navbar navbar-expand-md navbar-light bg-light">
-    <a class="navbar-brand" href="{{ url_for('home')}}">BU Fit</a>
+    <a class="navbar-brand" href="{{ url_for('main.home')}}">BU Fit</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
     </button>
     <div class="collapse navbar-collapse" id="navbarNav">
         <ul class="navbar-nav">
+            {% if current_user.is_authenticated %}
             <li class="nav-item active">
-                <a class="nav-link" href="{{ url_for('home')}}">Home</a>
+                <a class="nav-link" href="{{ url_for('main.home')}}">Home</a>
             </li>
             <li class="nav-item active">
-                <a class="nav-link" href="{{ url_for('search_exercises')}}">Search Exercises</a>
+                <a class="nav-link" href="{{ url_for('main.search_exercises')}}">Search Exercises</a>
             </li>
             <li class="nav-item active">
-                <a class="nav-link" href="{{ url_for('search_workouts')}}">Search Workouts</a>
+                <a class="nav-link" href="{{ url_for('main.search_workouts')}}">Search Workouts</a>
             </li>
             <li class="nav-item active">
-                <a class="nav-link" href="{{ url_for('library')}}">Library</a>
+                <a class="nav-link" href="{{ url_for('main.library')}}">Library</a>
             </li>
             <li class="nav-item active">
-                <a class="nav-link" href="{{ url_for('create_exercise')}}">Add Exercise</a>
+                <a class="nav-link" href="{{ url_for('main.create_exercise')}}">Add Exercise</a>
             </li>
             <li class="nav-item active">
-                <a class="nav-link" href="{{ url_for('create_workout')}}">Create Workout</a>
+                <a class="nav-link" href="{{ url_for('main.create_workout')}}">Create Workout</a>
             </li>
+            <li class="nav-item active">
+                <a class="nav-link" href="{{ url_for('auth.logout')}}">Log Out</a>
+            </li>
+            {% endif %}
+            {% if not current_user.is_authenticated %}
+            <li class="nav-item active">
+                <a class="nav-link" href="{{ url_for('auth.signup')}}">Sign Up</a>
+            </li>
+            {% endif %}
         </ul>
     </div>
 </nav>
