--- templates/create_account.html
@@ -1,24 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<h1>{% block title %} Create an Account {% endblock %}</h1>
-
-<form method="post">
-    <div class="form-group">
-        <label for='username'>Username</label>
-        <input type="text" name="username"
-               placeholder="username" class="form-control"
-               value="{{ request.form['username'] }}"></input>
-    </div>
-
-    <div class="form-group">
-        <label for='password'>Password</label>
-        <input type="text" name="password"
-               placeholder="password" class="form-control"
-               value="{{ request.form['password'] }}"></input>
-    </div>
-    <div class="form-group">
-        <button type="submit" class="btn btn-primary">Create</button>
-    </div>
-</form>
-{% endblock %}
