+++ templates/edit.html
@@ -31,7 +31,7 @@
 </form>
 <hr>
 
-<form action="{{ url_for('delete', id=exercise['id']) }}" method="POST">
+<form action="{{ url_for('main.delete', id=exercise['id']) }}" method="POST">
     <input type="submit" value="Delete Exercise"
            class="btn btn-danger btn-sm"
            onclick="return confirm('Are you sure you want to delete this exercise?')">
