+++ templates/login.html
@@ -1,27 +1,36 @@
-{% extends 'base.html' %}
+{% extends "base.html" %}
 
 {% block content %}
-<h1>{% block title %} Login Page {% endblock %}</h1>
+<div class="column is-4 is-offset-4">
+    <h3 class="title">Login</h3>
+    <div class="box">
+        {% with messages = get_flashed_messages() %}
+        {% if messages %}
+            <div class="notification is-danger">
+                {{ messages[0] }}
+            </div>
+        {% endif %}
+        {% endwith %}
+        <form method="POST" action="/login">
+            <div class="field">
+                <div class="control">
+                    <input class="input is-large" type="email" name="email" placeholder="Your Email" autofocus="">
+                </div>
+            </div>
 
-<form method="post">
-    <div class="form-group">
-        <label for='username'>Username</label>
-        <input type="text" name="username"
-               placeholder="username" class="form-control"
-               value="{{ request.form['username'] }}"></input>
+            <div class="field">
+                <div class="control">
+                    <input class="input is-large" type="password" name="password" placeholder="Your Password">
+                </div>
+            </div>
+            <div class="field">
+                <label class="checkbox">
+                    <input type="checkbox" name="remember">
+                    Remember me
+                </label>
+            </div>
+            <button class="button is-block is-info is-large is-fullwidth">Login</button>
+        </form>
     </div>
-
-    <div class="form-group">
-        <label for='password'>Password</label>
-        <input type="text" name="password"
-               placeholder="password" class="form-control"
-               value="{{ request.form['password'] }}"></input>
-    </div>
-    <div class="form-group">
-        <button type="submit" class="btn btn-primary">Login</button>
-    </div>
-</form>
-<a href="{{ url_for('create_account') }}">
-    <h2>Create Account</h2>
-</a>
+</div>
 {% endblock %}
