+++ templates/search_workouts.html
@@ -2,19 +2,19 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <link rel="stylesheet" href="{{ url_for('static', filename= 'css/style.css') }}">
+    <link rel="stylesheet" href="{{ url_for('main.static', filename= 'css/style.css') }}">
     <title>BUFit</title>
 </head>
 {% block content %}
 <h1>{% block title %} Workouts {% endblock %}</h1>
 {% for workout in workouts %}
-<a href="{{ url_for('workout', workout_id=workout['id']) }}">
+<a href="{{ url_for('main.workout', workout_id=workout['id']) }}">
     <h2>{{ workout['name'] }}</h2>
 </a>
 <p>{{ workout['author'] }}</p>
 <p>{{ workout['exercises'] }}</p>
 <span class="badge badge-primary">{{ workout['created'] }}</span>
-<a href="{{ url_for('edit_workout', workout_id=workout['id']) }}">
+<a href="{{ url_for('main.edit_workout', workout_id=workout['id']) }}">
     <span class="badge badge-warning">Edit</span>
 </a>
 <hr>
