+++ templates/workout.html
@@ -9,7 +9,7 @@
 <h2>Exercise List</h2>
 {% if selected_exercises %}
 {% for selected_exercise in selected_exercises %}
-    <a href="{{ url_for('exercise', exercise_id=selected_exercise['id']) }}">{{selected_exercise['name']}} </a>
+    <a href="{{ url_for('main.exercise', exercise_id=selected_exercise['id']) }}">{{selected_exercise['name']}} </a>
     <a>Sets: {{ selected_exercise['num_sets'] }}</a>
     <a>Reps: {{ selected_exercise['num_reps'] }}</a>
     <br>
